+++ NASAProj.Servicevvv/DTOs/CertificateForCreationDto.cs
@@ -1,12 +1,12 @@
-using ZaminEducation.Service.DTOs.UserCourses;
+using NASAProj.Service.DTOs.UserCourses;
 
 namespace NASAProj.Service.DTOs
 {
     public class CertificateForCreationDTO
     {
-        public long CourseId { get; set; }
+        public int CourseId { get; set; }
 
-        public long? UserId { get; set; }
+        public int? UserId { get; set; }
 
         public CertificateResultDTO Result { get; set; }
     }
